+++ Editor/AddPackages.cs
@@ -12,17 +12,13 @@ namespace JamStarter.Editor
     {
         private static readonly string PackagesDirectoryPath = Path.Join(Application.dataPath, "..","Packages");
         private static readonly string PackagesManifestPath = Path.Join(PackagesDirectoryPath, "manifest.json");
-    
-        private static readonly string[] Packages = {
-            "com.dbrizov.naughtyattributes",
-            "com.gitamend.unityutils",
-            "ayellowpaper.serialized-dictionary"
-        };
-    
-        private static readonly string[] PackageUrLs = {
-            "https://github.com/dbrizov/NaughtyAttributes.git#upm",
-            "https://github.com/adammyhre/Unity-Utils.git",
-            "https://github.com/ayellowpaper/SerializedDictionary.git"
+
+        private static readonly (string packageId, string packageUrl)[] Packages =
+        {
+            ("com.dbrizov.naughtyattributes", "https://github.com/dbrizov/NaughtyAttributes.git#upm"),
+            ("com.gitamend.unityutils", "https://github.com/adammyhre/Unity-Utils.git"),
+            ("ayellowpaper.serialized-dictionary", "https://github.com/ayellowpaper/SerializedDictionary.git"),
+            ("com.cysharp.unitask", "https://github.com/Cysharp/UniTask.git?path=src/UniTask/Assets/Plugins/UniTask"),
         };
     
         [InitializeOnLoadMethod]
@@ -37,7 +33,7 @@ namespace JamStarter.Editor
             var didChangeDependencies = false;
             for (var i = 0; i < Packages.Length; i++)
             {
-                var packageId = Packages[i];
+                var (packageId, packageUrl) = Packages[i];
                 if (dependencies.ContainsKey(packageId))
                 {
                     //Debug.Log($"Contains {packageId}");
@@ -45,7 +41,7 @@ namespace JamStarter.Editor
                 }
 
                 didChangeDependencies = true;
-                dependencies.Add(packageId, PackageUrLs[i]);
+                dependencies.Add(packageId, packageUrl);
                 //Debug.Log($"Added {packageId}");
             }